[#615] Fixes for PR comments
