+++ app/src/main/java/org/akvo/flow/util/logging/TagsFactory.java
@@ -38,7 +38,7 @@ import timber.log.Timber;
 
 public class TagsFactory {
 
-    private static final String INSTANCE_ID_TAG_KEY = "app.id";
+    private static final String GAE_INSTANCE_ID_TAG_KEY = "gae.instance";
     private static final String DEVICE_ID_TAG_KEY = "flow.device.id";
     private static final String DEVICE_MODEL_TAG_KEY = "device.model";
     private static final String OS_VERSION_TAG_KEY = "os.version";
@@ -59,7 +59,7 @@ public class TagsFactory {
 
     private void initTags(Context context) {
         tags.put(DEVICE_MODEL_TAG_KEY, Build.MODEL);
-        tags.put(INSTANCE_ID_TAG_KEY, getAppId(context));
+        tags.put(GAE_INSTANCE_ID_TAG_KEY, getAppId(context));
         tags.put(DEVICE_ID_TAG_KEY, getDeviceId(context));
         tags.put(OS_VERSION_TAG_KEY, Build.VERSION.RELEASE);
         try {
